# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Permionics BD Suite API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface AuthStatus {
  authenticated: boolean;
  /** @nullable */
  message?: string | null;
}

export interface LoginInput {
  /** @minLength 1 */
  password: string;
  remember?: boolean;
}

export interface ChangePasswordInput {
  /** @minLength 1 */
  currentPassword: string;
  /** @minLength 6 */
  newPassword: string;
}

export interface ActivityItem {
  type: string;
  id?: number;
  title: string;
  sector: string;
  createdAt: string;
}

export interface DashboardStats {
  totalCaseStudies: number;
  totalQuestionnaires: number;
  sectorsCount: number;
  /** @nullable */
  lastUpdated: string | null;
  recentActivity: ActivityItem[];
}

export interface CaseStudy {
  id: number;
  clientName: string;
  sector: string;
  location: string;
  challenge: string;
  solution: string;
  technologyStack: string;
  capacity: string;
  results: string;
  /** @nullable */
  testimonial?: string | null;
  tags: string[];
  fullText: string;
  createdAt: string;
  updatedAt: string;
}

export interface CaseStudyInput {
  /** @minLength 1 */
  clientName: string;
  /** @minLength 1 */
  sector: string;
  /** @minLength 1 */
  location: string;
  /** @minLength 1 */
  challenge: string;
  /** @minLength 1 */
  solution: string;
  /** @minLength 1 */
  technologyStack: string;
  /** @minLength 1 */
  capacity: string;
  /** @minLength 1 */
  results: string;
  /** @nullable */
  testimonial?: string | null;
  tags: string[];
  fullText: string;
}

export interface CaseStudyPatch {
  /** @minLength 1 */
  clientName?: string;
  sector?: string;
  location?: string;
  challenge?: string;
  solution?: string;
  technologyStack?: string;
  capacity?: string;
  results?: string;
  /** @nullable */
  testimonial?: string | null;
  tags?: string[];
  fullText?: string;
}

export interface AnswerMap {[key: string]: string}

export interface QuestionItem {
  id: string;
  question: string;
  type: string;
  options?: string[];
}

export interface Questionnaire {
  id: number;
  clientName: string;
  sector: string;
  questions: QuestionItem[];
  answers: AnswerMap;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface QuestionnaireInput {
  /** @minLength 1 */
  clientName: string;
  /** @minLength 1 */
  sector: string;
  questions: QuestionItem[];
  answers: AnswerMap;
  /** @nullable */
  notes?: string | null;
}

export interface ChatHistoryItem {
  role: string;
  content: string;
}

export interface ChatMessageInput {
  /** @minLength 1 */
  message: string;
  history: ChatHistoryItem[];
}

export interface ChatSource {
  id: number;
  clientName: string;
  sector: string;
  technologyStack?: string;
}

export interface ChatResponse {
  answer: string;
  sources: ChatSource[];
}

export type ListCaseStudiesParams = {
sector?: string;
technology?: string;
search?: string;
};