# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Permionics BD Suite API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /case-studies:
    get:
      operationId: listCaseStudies
      tags: [case-studies]
      summary: List all case studies with optional filters
      parameters:
        - name: sector
          in: query
          required: false
          schema:
            type: string
        - name: technology
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of case studies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseStudy"
    post:
      operationId: createCaseStudy
      tags: [case-studies]
      summary: Create a new case study
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseStudyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudy"
  /case-studies/{id}:
    get:
      operationId: getCaseStudy
      tags: [case-studies]
      summary: Get a single case study by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case study detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudy"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateCaseStudy
      tags: [case-studies]
      summary: Update an existing case study
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseStudyPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudy"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteCaseStudy
      tags: [case-studies]
      summary: Delete a case study
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
    AuthStatus:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean
        message:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 1
        remember:
          type: boolean
...
    CaseStudy:
      type: object
      required: [id, clientName, sector, location, challenge, solution, technologyStack, capacity, results, tags, fullText, createdAt, updatedAt]
      properties:
        id:
          type: integer
        clientName:
          type: string
        sector:
          type: string
        location:
          type: string
        challenge:
          type: string
        solution:
          type: string
        technologyStack:
          type: string
        capacity:
          type: string
        results:
          type: string
        testimonial:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        fullText:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    CaseStudyInput:
      type: object
      required: [clientName, sector, location, challenge, solution, technologyStack, capacity, results, tags, fullText]
      properties:
        clientName:
          type: string
          minLength: 1
        sector:
          type: string
          minLength: 1
        location:
          type: string
          minLength: 1
        challenge:
          type: string
          minLength: 1
        solution:
          type: string
          minLength: 1
        technologyStack:
          type: string
          minLength: 1
        capacity:
          type: string
          minLength: 1
        results:
          type: string
          minLength: 1
        testimonial:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        fullText:
          type: string
    CaseStudyPatch:
      type: object
      properties:
        clientName:
          type: string
          minLength: 1
        sector:
          type: string
        location:
          type: string
        challenge:
          type: string
        solution:
          type: string
        technologyStack:
          type: string
        capacity:
          type: string
        results:
          type: string
        testimonial:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        fullText:
          type: string
...